feat(1) refactor, return pdf
