--- src/main/java/ru/veselov/springstickers/SpringStickers/controllers/AuthController.java
@@ -1,17 +1,0 @@
-package ru.veselov.springstickers.SpringStickers.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AuthController {
-
-    @GetMapping("/login")
-    public String loginPage(){
-        return "login";
-    }
-
-
-
-
-}
