--- src/main/java/ru/veselov/springstickers/SpringStickers/model/SerialEntity.java
@@ -1,27 +1,0 @@
-package ru.veselov.springstickers.SpringStickers.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.util.Date;
-
-@Entity
-@Data
-@Table(name = "serial_num")
-public class SerialEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int serial_id;
-
-    @Column(name = "num")
-    private String num;
-
-    @Column(name="created_at")
-    private Date createdAt;
-
-    @ManyToOne
-    @JoinColumn(name="id", nullable = false)
-    private LabelEntity label;
-
-}
