--- src/main/java/ru/veselov/springstickers/SpringStickers/repositories/LabelRepository.java
@@ -1,13 +1,0 @@
-package ru.veselov.springstickers.SpringStickers.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.veselov.springstickers.SpringStickers.model.LabelEntity;
-
-import java.util.Optional;
-
-@Repository
-public interface LabelRepository extends JpaRepository<LabelEntity,Integer> {
-
-    Optional<LabelEntity> findByArticle(String article);
-}
