+++ src/main/java/ru/veselov/springstickers/springstickers/SpringStickersApplication.java
@@ -1,0 +1,21 @@
+package ru.veselov.springstickers.springstickers;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+@SpringBootApplication
+public class SpringStickersApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringStickersApplication.class, args);
+	}
+
+
+	@Bean
+	public ModelMapper modelMapper(){
+		return new ModelMapper();
+	}
+
+}
