+++ src/main/java/ru/veselov/springstickers/springstickers/config/SecurityConfig.java
@@ -1,0 +1,55 @@
+package ru.veselov.springstickers.springstickers.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import ru.veselov.springstickers.springstickers.services.PersonDetailsService;
+
+@EnableMethodSecurity
+@Configuration
+@RequiredArgsConstructor
+public class SecurityConfig {
+    public static final String LOGIN = "/login";
+    private final PersonDetailsService personDetailsService;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
+        httpSecurity.authorizeHttpRequests(
+                        r -> r.requestMatchers("/", "/show", "/download", LOGIN).permitAll()
+                                .requestMatchers("/admin", "/admin/add", "/admin/serials", "/admin/manage",
+                                        "/admin/edit/**", "/admin/delete/**").hasAnyRole("ADMIN"))
+                .formLogin(login -> login.loginProcessingUrl("/process")
+                        .defaultSuccessUrl("/admin")
+                        .failureUrl("/login?error")
+                        .loginPage(LOGIN))
+                .logout(logout -> logout.logoutSuccessUrl(LOGIN)
+                        .logoutUrl(("/logout")));
+        return httpSecurity.build();
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+        return configuration.getAuthenticationManager();
+    }
+
+    @Bean
+    public DaoAuthenticationProvider authenticationProvider() {
+        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
+        authenticationProvider.setUserDetailsService(personDetailsService);
+        authenticationProvider.setPasswordEncoder(bCryptPasswordEncoder());
+        return authenticationProvider;
+    }
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+}
