+++ src/main/java/ru/veselov/springstickers/springstickers/controllers/AuthController.java
@@ -1,0 +1,17 @@
+package ru.veselov.springstickers.springstickers.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AuthController {
+
+    @GetMapping("/login")
+    public String loginPage(){
+        return "login";
+    }
+
+
+
+
+}
