+++ src/main/java/ru/veselov/springstickers/springstickers/exception/PdfCreationException.java
@@ -1,0 +1,8 @@
+package ru.veselov.springstickers.springstickers.exception;
+
+public class PdfCreationException extends RuntimeException {
+    public PdfCreationException(String message) {
+        super(message);
+    }
+
+}
