+++ src/main/java/ru/veselov/springstickers/springstickers/services/SaveToFileService.java
@@ -1,0 +1,11 @@
+package ru.veselov.springstickers.springstickers.services;
+
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+
+public interface SaveToFileService {
+
+    InputStream saveToPdf(ByteArrayOutputStream imageStream);
+
+}
