+++ src/test/java/ru/veselov/springstickers/SpringControllerTest.java
@@ -14,13 +14,13 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import ru.veselov.springstickers.SpringStickers.SpringStickersApplication;
-import ru.veselov.springstickers.SpringStickers.services.SerialService;
-import ru.veselov.springstickers.SpringStickers.controllers.SpringLabelController;
-import ru.veselov.springstickers.SpringStickers.model.LabelEntity;
-import ru.veselov.springstickers.SpringStickers.model.LabelSticker;
-import ru.veselov.springstickers.SpringStickers.services.LabelService;
-import ru.veselov.springstickers.SpringStickers.services.PaperService;
+import ru.veselov.springstickers.springstickers.SpringStickersApplication;
+import ru.veselov.springstickers.springstickers.services.SerialService;
+import ru.veselov.springstickers.springstickers.controllers.SpringLabelController;
+import ru.veselov.springstickers.springstickers.entity.LabelEntity;
+import ru.veselov.springstickers.springstickers.model.LabelSticker;
+import ru.veselov.springstickers.springstickers.services.LabelService;
+import ru.veselov.springstickers.springstickers.services.PaperService;
 
 import java.io.ByteArrayInputStream;
 import java.util.HashMap;
